refactor(javm-cap): promote talc cap shape to crate root; legacy moves to legacy/
Commit 1 of the "consolidate to a single Cap type" plan
(/home/jar/.claude/plans/distributed-puzzling-tower.md).

The talc-shape Cap, CNode, Cache, etc. live at `javm_cap::*` directly.
The legacy identity-only Cap moves intact to `javm_cap::legacy::*` for
the duration of the migration; consumers in jar-kernel/javm rewrite
their imports to `use javm_cap::legacy::{...}` with no semantic
change. The talc/ subdirectory and the parallel `pub mod talc`
namespace go away.

File moves (preserving git history via `git mv`):
- talc/{cap,cnode,data,instance,page,entry,cache}.rs → crate root
- talc/hash.rs → cap_hash.rs (the root hash.rs is the Blake2b256 primitive)
- talc/image.rs → image_cap.rs (the root image.rs is the SCALE Image)
- talc/{cache_tests.rs, tests.rs} → cache_tests.rs / cap_tests.rs at root
- (old) cap.rs, cnode.rs, ops.rs → legacy/{cap,cnode,ops}.rs
- new legacy/mod.rs aggregates the re-exports

API additions on the new javm_cap::Cap:
- derive(Clone, Debug) on Cap<A>, InstanceCap<A>, ImageCap<A>,
  DataCap<A>, CNodeCap<A>, RwOverlay<A>, DataContent<A>, PageSlot<A>,
  PageBytes<A> — and Aarc<T, A> in nub-host-common (required for
  PageSlot's Loaded variant to derive Debug)
- CNodeCap<A>::set / take / remove — binary-search insert/remove
  keeping `slots` sorted; rejects out-of-range slot indices
- CNodeCap<A>::new_in now returns Result<Self, CapError> (rejects
  size_log > 16 matching the legacy InMemoryCNode::new contract)
- CNodeCap<Global>::new(size_log) — heap convenience alias
- Cap::data_inline(bytes) / Cap::image_from(&Image) / Cap::empty_cnode(size_log)
  — Global-only heap constructors; jar-kernel and javm use them to
  build caps without going through a Cache

Workspace + clippy + all 606 tests stay green. nub smoke still passes
both backends.
