+++ rust/javm-cap/src/entry.rs
