+++ rust/javm/src/frame.rs
@@ -26,7 +26,9 @@
 //! `Cap::CNode` slots) is provided here too: the MGMT dispatcher
 //! takes a `SlotPath` operand, this module walks it.
 
-use javm_cap::{CNodeBackend, Cap, CapError, PinnedCap, SlotIdx, SlotPath, image::Image};
+use javm_cap::image::{Image, PinnedCap};
+use javm_cap::legacy::{CNodeBackend, Cap};
+use javm_cap::{CapError, SlotIdx, SlotPath};
 
 use crate::error::VmError;
 
@@ -121,7 +123,7 @@ impl<'a> BareFrame<'a> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use javm_cap::{CNodeCap, InMemoryCNode};
+    use javm_cap::legacy::{CNodeCap, InMemoryCNode};
     use std::collections::BTreeMap;
     use std::sync::Arc;
 
@@ -150,7 +152,7 @@ mod tests {
 
     #[test]
     fn mainframe_get_root_slot() {
-        let cap = Cap::Image(javm_cap::ImageCap {
+        let cap = Cap::Image(javm_cap::legacy::ImageCap {
             content_hash: [1; 32],
         });
         let cn = cnode_with(&[(SlotIdx(5), cap.clone())]);
@@ -178,7 +180,7 @@ mod tests {
 
     #[test]
     fn mainframe_resolve_root_path() {
-        let cap = Cap::Image(javm_cap::ImageCap {
+        let cap = Cap::Image(javm_cap::legacy::ImageCap {
             content_hash: [7; 32],
         });
         let cn = cnode_with(&[(SlotIdx(9), cap)]);
@@ -194,7 +196,7 @@ mod tests {
         // Inner cnode at slot 3 of root; inner slot 1 holds a Cap::Type.
         let inner = cnode_with(&[(
             SlotIdx(1),
-            Cap::Type(javm_cap::TypeCap {
+            Cap::Type(javm_cap::legacy::TypeCap {
                 image_hash_chain: [42; 32],
             }),
         )]);
@@ -211,7 +213,7 @@ mod tests {
     fn mainframe_resolve_non_cnode_intermediate_errors() {
         // Root slot 4 holds a Cap::Image (not a Cap::CNode); walking
         // through it should fail.
-        let img_cap = Cap::Image(javm_cap::ImageCap {
+        let img_cap = Cap::Image(javm_cap::legacy::ImageCap {
             content_hash: [9; 32],
         });
         let root = cnode_with(&[(SlotIdx(4), img_cap)]);
@@ -234,14 +236,14 @@ mod tests {
 
     #[test]
     fn bareframe_only_reads_pinned() {
-        let cap = Cap::Image(javm_cap::ImageCap {
+        let cap = Cap::Image(javm_cap::legacy::ImageCap {
             content_hash: [3; 32],
         });
         let cn = cnode_with(&[
             (SlotIdx(2), cap),
             (
                 SlotIdx(3),
-                Cap::Type(javm_cap::TypeCap {
+                Cap::Type(javm_cap::legacy::TypeCap {
                     image_hash_chain: [0; 32],
                 }),
             ),
